DataRequestHandler architecture for GetStatusHandler
