+++ src/main/java/org/icatproject/ids/services/RequestHandlerService.java
@@ -30,8 +30,6 @@ import org.icatproject.ids.requestHandlers.getDataFileIdsHandlers.GetDataFileIds
 import org.icatproject.ids.requestHandlers.getDataFileIdsHandlers.GetDataFileIdsHandlerForUnpreparedData;
 import org.icatproject.ids.requestHandlers.getSizeHandlers.GetSizeHandlerForPreparedData;
 import org.icatproject.ids.requestHandlers.getSizeHandlers.GetSizeHandlerForUnpreparedData;
-import org.icatproject.ids.requestHandlers.getStatusHandlers.GetStatusHandlerForPreparedData;
-import org.icatproject.ids.requestHandlers.getStatusHandlers.GetStatusHandlerForUnpreparedData;
 import org.icatproject.ids.requestHandlers.resetHandlers.ResetHandlerForPreparedData;
 import org.icatproject.ids.requestHandlers.resetHandlers.ResetHandlerForUnpreparedData;
 import org.slf4j.Logger;
@@ -158,9 +156,6 @@ public class RequestHandlerService {
 
                 this.registerHandler(new GetSizeHandlerForPreparedData());
                 this.registerHandler(new GetSizeHandlerForUnpreparedData());
-
-                this.registerHandler(new GetStatusHandlerForPreparedData());
-                this.registerHandler(new GetStatusHandlerForUnpreparedData());
 
                 this.registerHandler(new ResetHandlerForPreparedData());
                 this.registerHandler(new ResetHandlerForUnpreparedData());
